[codex] 通信byte列をuint8_tへ統一

+++ tests/parser/parser_test_input.c
@@ -1,6 +1,7 @@
 #include <assert.h>
 #include <stddef.h>
 #include <stdint.h>
+#include "lidar/core.h"
 #include "parser_test_input.h"
 
 static const uint8_t* test_input;
@@ -18,10 +19,10 @@ void parser_test_set_input(const uint8_t* input, size_t length)
 size_t parser_test_bytes_read(void) { return test_read_index; }
 
 /* parserが使うRXリングバッファ読み出し処理をホストテスト用に置き換える。 */
-int8_t read_byte(void)
+uint8_t read_byte(void)
 {
         assert(test_read_index < test_input_length);
-        return (int8_t)test_input[test_read_index++];
+        return test_input[test_read_index++];
 }
 
 /* バイト列のデコード方法をターゲット実装と同じにする。 */
@@ -30,7 +31,7 @@ uint32_t dec_little_endian(const uint8_t length)
         uint32_t value = 0u;
 
         for (uint8_t i = 0u; i < length; ++i)
-                value |= (uint32_t)(uint8_t)read_byte() << (i * 8u);
+                value |= (uint32_t)read_byte() << (i * 8u);
 
         return value;
 }
